create case module
